Struct numext_fixed_hash::H256

pub struct H256(pub [u8; 32]);

Fixed hash type.

Notice:

Avoid to use the inner type directly.

Methods

impl H256

pub fn repeat_byte(byte: u8) -> Self

Return a new fixed hash where all bytes are set to the given byte.

pub fn zero() -> Self

Create a new fixed hash and all bits of it are zeros.

pub fn is_zero(&self) -> bool

Test if all bits of a fixed hash are zero.

pub fn is_max(&self) -> bool

Test if all bits of a fixed hash are one.

pub fn covers(&self, hash: &Self) -> bool

Test if all bits set in a hash are also set in self.

pub fn count_bits() -> u64

Return the count of bits.

pub fn bit(&self, index: usize) -> Option<bool>

Return a specific bit, or return None when overlows.

pub fn set_bit(&mut self, index: usize, value: bool) -> bool

Set a specific bit. Return false when overflows.

pub fn highest_one(&self) -> Option<usize>

Return the highest bit which is one.

pub fn lowest_one(&self) -> Option<usize>

Return the lowest bit which is one.

pub fn count_bytes() -> u64

Return the count of bytes.

pub fn byte(&self, index: usize) -> Option<u8>

Return a specific byte, or return None when overlows.

pub fn set_byte(&mut self, index: usize, byte: u8) -> bool

Set a specific byte. Return false when overflows;

pub fn highest_nonzero_byte(&self) -> Option<usize>

Return the highest byte which is nonzero.

pub fn lowest_nonzero_byte(&self) -> Option<usize>

Return the lowest byte which is nonzero.

pub fn as_bytes(&self) -> &[u8]

Get the inner bytes slice of a fixed hash.

pub fn as_bytes_mut(&mut self) -> &mut [u8]

Get the mutable inner bytes slice of a fixed hash.

pub fn as_fixed_bytes(&self) -> &[u8; 32]

Get the inner bytes of a fixed hash.

pub fn as_fixed_bytes_mut(&mut self) -> &[u8; 32]

Get the mutable inner bytes of a fixed hash.

pub fn to_fixed_bytes(self) -> [u8; 32]

Get the inner bytes array of a fixed hash.

pub fn as_ptr(&self) -> *const u8

Get a constant raw pointer to the inner bytes array of a fixed hash.

pub fn as_mut_ptr(&mut self) -> *mut u8

Get a mutable raw pointer to the inner bytes array of a fixed hash.

pub fn from_slice(input: &[u8]) -> Result<Self, FixedHashError>

Convert from slice.

pub fn into_slice(&self, output: &mut [u8]) -> Result<(), FixedHashError>

Convert into slice.

pub fn from_hex_str(input: &str) -> Result<Self, FixedHashError>

Convert from a hexadecimal string.

pub fn min_value() -> Self

Returns the smallest value that can be represented by this integer type.

pub fn max_value() -> Self

Returns the largest value that can be represented by this integer type.

pub fn count_ones(&self) -> u32

Returns the number of ones in the binary representation of self.

pub fn count_zeros(&self) -> u32

Returns the number of zeros in the binary representation of self.

pub fn leading_zeros(&self) -> u32

Returns the number of leading zeros in the binary representation of self.

pub fn trailing_zeros(&self) -> u32

Returns the number of trailing zeros in the binary representation of self.

pub fn checked_shl(&self, rhs: u128) -> Option<Self>

Checked shift left. Computes self << rhs, returning None if rhs is larger than or equal to the number of bits in self.

pub fn checked_shr(&self, rhs: u128) -> Option<Self>

Checked shift right. Computes self >> rhs, returning None if rhs is larger than or equal to the number of bits in self.

pub fn checked_neg(&self) -> Option<Self>

Checked negation. Computes -self, returning None unless self == 0. Note that negating any positive integer will overflow.

pub fn overflowing_shl(&self, rhs: u128) -> (Self, bool)

Shifts self left by rhs bits.

Returns a tuple of the shifted version of self along with a boolean indicating whether the shift value was larger than or equal to the number of bits. If the shift value is too large, then value is masked (N-1) where N is the number of bits, and this value is then used to perform the shift.

pub fn overflowing_shr(&self, rhs: u128) -> (Self, bool)

Shifts self right by rhs bits.

Returns a tuple of the shifted version of self along with a boolean indicating whether the shift value was larger than or equal to the number of bits. If the shift value is too large, then value is masked (N-1) where N is the number of bits, and this value is then used to perform the shift.

pub fn random<R: RngCore>(rng: &mut R) -> Self

Create a random fixed uint with a input random core.

pub fn thread_random() -> Self

Create a random fixed uint.